change in maclGroups()

+++ admin/acl.php
@@ -291,11 +291,12 @@ function maclGroups()
 			}
 		}
 
-	foreach($_POST['tablelist'] as $table)
-		{
-		if (!isset($_POST['group'][$table]))
-			$db->db_query("DELETE FROM ".$table." WHERE id_object='".$id_object."' AND id_group < '".BAB_ACL_GROUP_TREE."'");
-		}
+	if (isset($_POST['tablelist']))
+		foreach($_POST['tablelist'] as $table)
+			{
+			if (!isset($_POST['group'][$table]))
+				$db->db_query("DELETE FROM ".$table." WHERE id_object='".$id_object."' AND id_group < '".BAB_ACL_GROUP_TREE."'");
+			}
 
 	
 	if (isset($_POST['tree']) && count($_POST['tree']) > 0) {
